requirements.txt
# Window Restore - Windows Window Position Preset Manager
# Requirements

pywin32>=306
psutil>=5.9.0
websocket-client>=1.6.0
pystray>=0.19.5
Pillow>=10.0.0
comtypes>=1.3.0
